Move integration event dispatcher interface to infrastructure layer
